test: extract AbstractFilterTest
