+++ tests/Authentication/Filters/TokenFilterTest.php
@@ -2,64 +2,21 @@
 
 namespace Tests\Authentication\Filters;
 
-use CodeIgniter\Config\Factories;
 use CodeIgniter\Shield\Entities\AccessToken;
 use CodeIgniter\Shield\Entities\User;
 use CodeIgniter\Shield\Filters\TokenAuth;
 use CodeIgniter\Shield\Models\UserModel;
-use CodeIgniter\Test\FeatureTestTrait;
-use Config\Services;
-use Tests\Support\DatabaseTestCase;
+use CodeIgniter\Test\DatabaseTestTrait;
 
 /**
  * @internal
  */
-final class TokenFilterTest extends DatabaseTestCase
+final class TokenFilterTest extends AbstractFilterTest
 {
-    use FeatureTestTrait;
+    use DatabaseTestTrait;
 
-    protected $namespace;
-
-    protected function setUp(): void
-    {
-        $_SESSION = [];
-
-        Services::reset(true);
-
-        parent::setUp();
-
-        // Register our filter
-        $this->registerFilter();
-
-        // Add a test route that we can visit to trigger.
-        $this->addRoutes();
-    }
-
-    private function registerFilter(): void
-    {
-        $filterConfig = config('Filters');
-
-        $filterConfig->aliases['tokenAuth'] = TokenAuth::class;
-
-        Factories::injectMock('filters', 'filters', $filterConfig);
-    }
-
-    private function addRoutes(): void
-    {
-        $routes = service('routes');
-
-        $routes->group('/', ['filter' => 'tokenAuth'], static function ($routes): void {
-            $routes->get('protected-route', static function (): void {
-                echo 'Protected';
-            });
-        });
-        $routes->get('open-route', static function (): void {
-            echo 'Open';
-        });
-        $routes->get('login', 'AuthController::login', ['as' => 'login']);
-
-        Services::injectMock('routes', $routes);
-    }
+    protected string $alias     = 'tokenAuth';
+    protected string $classname = TokenAuth::class;
 
     public function testFilterNotAuthorized(): void
     {
